+++ docs/changes.rst
@@ -6,6 +6,9 @@ Changelog
 
 - Upgrade plone to 5.2.9
 - Fix the way ``get_survey_templates`` handles a survey with an unset category.
+- Remove the ghost module and its tests
+  because it is not used since quite some time
+  [ale-rt]
 
 
 14.1.5 (2022-07-13)